+++ libcnb-test/tests/integration_test.rs
@@ -68,7 +68,7 @@ fn starting_containers() {
                     .env("PORT", TEST_PORT.to_string())
                     .expose_port(TEST_PORT),
                 |container| {
-                    let address_on_host = container.address_for_port(TEST_PORT).unwrap();
+                    let address_on_host = container.address_for_port(TEST_PORT);
                     let url = format!("http://{}:{}", address_on_host.ip(), address_on_host.port());
 
                     // Retries needed since the server takes a moment to start up.